refactored gui.cursor module
